addressed CR comments
