+++ blurr/cli/cli.py
@@ -1,6 +1,7 @@
 from blurr.cli.validate import validate_command
+from blurr.util.out import Out
 
 
-def cli(arguments, out):
-    if arguments["validate"]:
-        return validate_command(arguments["<DTC>"], out)
+def cli(arguments: list, out: Out) -> int:
+    if arguments['validate']:
+        return validate_command(arguments['<DTC>'], out)
