+++ blurr/cli/validate.py
@@ -2,20 +2,21 @@ import yaml
 
 from blurr.core.errors import InvalidSchemaError
 from blurr.core.syntax.schema_validator import validate
+from blurr.util.out import Out
 
 
-def validate_command(dtc_file, out):
+def validate_command(dtc_file: str, out: Out) -> int:
     try:
         dtc_dict = yaml.safe_load(open(dtc_file))
         validate(dtc_dict)
-        out.print("document is valid")
+        out.print('document is valid')
         return 0
     except yaml.YAMLError:
-        out.eprint("invalid yaml")
+        out.eprint('invalid yaml')
         return 1
     except InvalidSchemaError as err:
         out.eprint(str(err))
         return 1
     except:
-        out.eprint("there was an error parsing the document")
+        out.eprint('there was an error parsing the document')
         return 1
